Reboot KangarooTwelve.jl with a rudimentary pure-Julia implementation.
This commit reboots the KangarooTwelve.jl repository with a rudimentary
pure-Julia implementation. Specifically, apart from structuring the
repository, this commit checks in a transliteration of XKCP's Python
implementation of KangarooTwelve and the associated test suite.

The transliteration was fairly strict, including a minimal set of
modifications to make the test suite produce byte-identical results
to the test suite of the Python version. That is to say that, like
XKCP's implementation and test suite, the code checked in here
comes complete with obscure identifier names, a lack of comments,
and test reliance upon visual inspection for longitudinal consistency.
Additionally, while the transliteration seems to be roughly a factor
of four faster than the original Python code (Python 3.8 / Julia 1.4),
it seems to be roughly three orders of magnitude slower than the
C library formerly wrapped in this repository.

In other words, this code could use some love and care prior to deployment :).